Move abstract_sensorimotor_test.py to test directory
